+++ code/nmt-backend-basic/fpbasic-module-nmt/src/main/java/cn/fengp/basic/module/nmt/service/courseinfo/CourseInfoService.java
@@ -20,7 +20,7 @@ public interface CourseInfoService {
      * @param createReqVO 创建信息
      * @return 编号
      */
-    Integer createCourseInfo(@Valid CourseInfoSaveReqVO createReqVO);
+    Long createCourseInfo(@Valid CourseInfoSaveReqVO createReqVO);
 
     /**
      * 更新课程信息
@@ -34,14 +34,14 @@ public interface CourseInfoService {
      *
      * @param id 编号
      */
-    void deleteCourseInfo(Integer id);
+    void deleteCourseInfo(Long id);
 
     /**
     * 批量删除课程信息
     *
     * @param ids 编号
     */
-    void deleteCourseInfoListByIds(List<Integer> ids);
+    void deleteCourseInfoListByIds(List<Long> ids);
 
     /**
      * 获得课程信息
@@ -49,7 +49,7 @@ public interface CourseInfoService {
      * @param id 编号
      * @return 课程信息
      */
-    CourseInfoDO getCourseInfo(Integer id);
+    CourseInfoDO getCourseInfo(Long id);
 
     /**
      * 获得课程信息分页
